+++ R/imports.R
@@ -1,0 +1,4 @@
+#' @importFrom FNN get.knnx
+#' @import methods
+#' @noRd
+"_PACKAGE"